Use an int for the maximum object size to be cached in bytes since it is cached using a byte array and that is limited to Integer.MAX_VALUE. Test objects to be cached against the maximum object size, not the maximum cache size.
git-svn-id: https://svn.apache.org/repos/asf/tomcat/trunk@1549706 13f79535-47bb-0310-9956-ffa450edef68

+++ java/org/apache/catalina/webresources/StandardRoot.java
@@ -455,12 +455,12 @@ public class StandardRoot extends LifecycleMBeanBase
     }
 
     @Override
-    public void setCacheMaxObjectSize(long cacheMaxObjectSize) {
+    public void setCacheMaxObjectSize(int cacheMaxObjectSize) {
         cache.setMaxObjectSize(cacheMaxObjectSize);
     }
 
     @Override
-    public long getCacheMaxObjectSize() {
+    public int getCacheMaxObjectSize() {
         return cache.getMaxObjectSize();
     }
